rename compileToCSS function
